requirements.txt
redis
fastapi
python-dotenv 
httpx
fastapi[standard]
slowapi
